Move SegmentUsingEllipsoidsView to UI package
