+++ src/main/java/org/mastodon/mamut/segment/SegmentByEllipsoidsPlugin.java
@@ -8,6 +8,7 @@ import org.mastodon.app.ui.ViewMenuBuilder;
 import org.mastodon.mamut.MamutAppModel;
 import org.mastodon.mamut.plugin.MamutPlugin;
 import org.mastodon.mamut.plugin.MamutPluginAppModel;
+import org.mastodon.mamut.segment.ui.SegmentUsingEllipsoidsView;
 import org.scijava.Context;
 import org.scijava.command.CommandService;
 import org.scijava.plugin.Parameter;
